🐛 Fix wrong permission bug
