+++ src/App.js
@@ -13,7 +13,7 @@ const Manage = React.lazy(() => import('./pages/Manage/Manage'));
 const Register = React.lazy(() => import('./pages/Register/Register'));
 const Sample = React.lazy(() => import('./pages/Sample/Sample'));
 const SignDocument = React.lazy(() => import('./pages/SignDocument/SignDocument'));
-const Signing = React.lazy(() => import('./pages/Signing/Signing2'));
+const Signing = React.lazy(() => import('./pages/Signing/Signing'));
 const PrivateRoute = React.lazy(() => import('./components/Layout/PrivateRoute'));
 const Activation = React.lazy(() => import('./pages/Activation/Activation'));
 const UserProfile = React.lazy(() => import('./pages/UserProfile/UserProfile'));
